% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_arulesViz-package.R
\docType{package}
\name{arulesViz-package}
\alias{arulesViz-package}
\title{arulesViz: Visualizing Association Rules and Frequent Itemsets}
\description{
Extends package 'arules' with various visualization techniques for association rules and itemsets. The package also includes several interactive visualizations for rule exploration. Michael Hahsler (2017) .
}
\section{Main Functions}{

\itemize{
\item \code{\link[=plot]{plot()}} plot association rules.
\item \code{\link[=inspectDT]{inspectDT()}} for interactive rule tables.
\item \code{\link[=ruleExplorer]{ruleExplorer()}} for interactive rule inspection with tables and graphs.
}
}

\references{
Hahsler M (2017). arulesViz: Interactive Visualization of Association Rules with R. R Journal, 9(2), 163–175. ISSN 2073-4859, \doi{10.32614/RJ-2017-047}.
}
\author{
Michael Hahsler
}
