% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules2groupedmatrix.R, R/rules2matrix.R
\name{rules2groupedMatrix}
\alias{rules2groupedMatrix}
\alias{rules2matrix}
\title{Convert association rules into a matrix}
\usage{
rules2groupedMatrix(
  rules,
  measure = "lift",
  measure2 = "support",
  k = 10,
  aggr.fun = mean,
  lhs_label_items = 2
)

rules2matrix(rules, measure = "support", reorder = "measure", ...)
}
\arguments{
\item{rules}{a rules object.}

\item{measure}{quality measure put in the matrix}

\item{measure2}{second quality measure (organized in the same way as
measure).}

\item{k}{number of LHS itemset groups.}

\item{aggr.fun}{function to aggregate the quality measure for groups.}

\item{lhs_label_items}{number of top items used to name LHS itemset groups
(columns).}

\item{reorder}{reorder rows and columns? Possible methods are: "none",
"measure" (default), "support/confidence", "similarity".}

\item{...}{passed on to \code{\link[arules:DATAFRAME]{arules::DATAFRAME()}}.}
}
\value{
\code{rules2matrix} returns a matrix with quality values.

\code{rules2groupedMatrix} returns a list with elements
\item{m}{ the
grouped matrix for measure. }
\item{m2}{ the grouped matrix for measure2. }
\item{clustering_rules}{ vector with group assignment for each rule. }
}
\description{
Converts a set of association rules into a matrix with unique LHS itemsets
as columns and unique RHS itemsets as rows. The matrix cells contain a
quality measure. The LHS itemsets can be grouped.
}
\examples{

data(Groceries)
rules <- apriori(Groceries, parameter=list(support = 0.001, confidence = 0.8))
rules

## Matrix
m <- rules2matrix(rules[1:10], measure = "lift")
m
plot(rules[1:10], method = "matrix")

## Grouped matrix
# create a matrix with LHSs grouped in k = 10 groups
gm <- rules2groupedMatrix(rules, k = 10)
gm$m

# number of rules per group 
table(gm$clustering_rules)

# get rules for group 1
inspect(rules[gm$clustering_rules == 1])

# create the corresponding grouped matrix plot by passing the grouped matrix as the groups parameter
plot(rules, method = "grouped matrix", groups = gm)

}
\references{
Michael Hahsler and Radoslaw Karpienko. Visualizing association
rules in hierarchical groups. Journal of Business Economics, 87(3):317--335,
May 2016. \doi{10.1007/s11573-016-0822-8}.
}
\seealso{
\code{\link{plot}} for rules using \code{method = 'matrix'} and
\code{method = 'grouped matrix'}.
}
\author{
Michael Hahsler
}
